+++ turbo/apps/app-worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.8.37](https://github.com/vm0-ai/okou/compare/app-worker-v1.8.36...app-worker-v1.8.37) (2026-09-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/core bumped to 8.670.0
+
 ## [1.8.36](https://github.com/vm0-ai/okou/compare/app-worker-v1.8.35...app-worker-v1.8.36) (2026-09-16)